refactor resultsScrollBar
