+++ app/frontend/src/classes/Results/ResultsScrollBar.ts
@@ -1,12 +1,12 @@
-import { storeManager } from '../store/StoreManager';
-import { TR_HEIGHT } from '../global.js';
+import { storeManager } from '../../store/StoreManager';
+import { TR_HEIGHT } from '../../global.js';
 
 // ================================================================
 // CONSTANTS & TYPES
 // ================================================================
 
 const RELEASE_DURATION = 2000;
-const MIN_HEIGHT = 30;
+const MIN_SCROLLBAR_HEIGHT = 30;
 
 /**
  * Represents the state of a drag operation
@@ -26,25 +26,51 @@ interface DragEventUI {
   };
 }
 
+/**
+ * Scroll calculation result type
+ */
+type ScrollCalculation = {
+  totalHeight: number;
+  availableScrollY: number;
+  newScrollPosition: number;
+};
+
+/**
+ * ScrollBar calculation result type
+ */
+type ScrollBarCalculation = {
+  barHeight: number;
+  barTop: number;
+  displayRate: number;
+};
+
+/**
+ * Scroll bounds type
+ */
+type ScrollBounds = {
+  min: number;
+  max: number;
+};
+
 // ================================================================
 // MAIN CLASS
 // ================================================================
 
 /**
- * Custom scrollbar component for large datasets with drag functionality
+ * Results専用スクロールバーコンポーネント
  *
  * ## Features
- * - Mouse drag support for desktop devices
- * - Touch drag support for mobile devices
- * - Automatic position calculation based on data
- * - Visual feedback during interactions
+ * - Results表示に特化したスクロール処理
+ * - マウス・タッチドラッグ対応
+ * - 自動位置計算とフィードバック
+ * - パフォーマンス最適化済み
  *
  * ## Usage
  * ```typescript
- * const scrollbar = new ScrollBar(containerElement);
+ * const scrollbar = new ResultsScrollBar(containerElement);
  * ```
  */
-export default class ScrollBar {
+export default class ResultsScrollBar {
   // ================================================================
   // PROPERTIES
   // ================================================================
@@ -68,13 +94,15 @@ export default class ScrollBar {
   private touchStartY: number = 0;
   /** Initial top position for touch drag */
   private touchStartTop: number = 0;
+  /** Last scroll position for delta calculations */
+  private lastScrollPosition: number = 0;
 
   // ================================================================
   // CONSTRUCTOR & INITIALIZATION
   // ================================================================
 
   /**
-   * Creates a new ScrollBar instance
+   * Creates a new ResultsScrollBar instance
    * @param containerElement - The container element where the scrollbar will be inserted
    */
   constructor(containerElement: HTMLElement) {
@@ -206,6 +234,66 @@ export default class ScrollBar {
   }
 
   // ================================================================
+  // PUBLIC API - Results-specific methods
+  // ================================================================
+
+  /**
+   * Deactivate the scrollbar visual state
+   */
+  deactivate(): void {
+    this.container.classList.remove('-active');
+  }
+
+  /**
+   * Initialize scrollbar position
+   */
+  initializePosition(): void {
+    this.container.classList.add('-active');
+  }
+
+  /**
+   * Handle scroll with scrollbar feedback (for touch events)
+   * @param deltaY - Y delta value
+   * @param touchStartOffset - Starting offset when touch began
+   */
+  handleScrollWithFeedback(deltaY: number, touchStartOffset: number): void {
+    const newOffset = this._calculateTouchScrollOffset(
+      deltaY,
+      touchStartOffset
+    );
+    const boundedOffset = this._clampOffsetToValidRange(newOffset);
+
+    this.lastScrollPosition = boundedOffset * TR_HEIGHT;
+    storeManager.setData('offset', boundedOffset);
+    this.updateDirectly(boundedOffset);
+  }
+
+  /**
+   * Handle simple scroll (for wheel events)
+   * @param deltaY - Y delta value
+   */
+  handleScroll(deltaY: number): void {
+    const calculation = this._calculateScrollPosition(deltaY);
+
+    if (calculation.newScrollPosition === this.lastScrollPosition) {
+      return;
+    }
+
+    this.lastScrollPosition = calculation.newScrollPosition;
+    const offset = this._calculateOffsetFromScroll(this.lastScrollPosition);
+    storeManager.setData('offset', offset);
+  }
+
+  /**
+   * Update scrollbar directly with specific offset
+   * @param offset - Offset value
+   */
+  updateDirectly(offset: number): void {
+    const calculation = this._calculateScrollBarPosition(offset);
+    this._applyScrollBarStyles(this.scrollBarElement, calculation, offset);
+  }
+
+  // ================================================================
   // PRIVATE - Visual Updates
   // ================================================================
 
@@ -223,7 +311,8 @@ export default class ScrollBar {
     const displayRate = displayHeight / totalHeight;
 
     let barHeight = Math.ceil(displayHeight * displayRate);
-    barHeight = barHeight < MIN_HEIGHT ? MIN_HEIGHT : barHeight;
+    barHeight =
+      barHeight < MIN_SCROLLBAR_HEIGHT ? MIN_SCROLLBAR_HEIGHT : barHeight;
 
     const availableHeight = displayHeight - barHeight * 0;
     const availableRate = availableHeight / totalHeight;
@@ -259,6 +348,130 @@ export default class ScrollBar {
    */
   private _releaseVisualState(): void {
     this.container.classList.remove('-dragging');
+  }
+
+  // ================================================================
+  // PRIVATE - Scroll Calculations (integrated from ScrollHandler)
+  // ================================================================
+
+  /**
+   * Calculate scroll position
+   */
+  private _calculateScrollPosition(deltaY: number): ScrollCalculation {
+    const numberOfRecords = storeManager.getData('numberOfRecords') as number;
+    const rowCount = storeManager.getData('rowCount') as number;
+
+    const totalHeight = numberOfRecords * TR_HEIGHT;
+    let availableScrollY = totalHeight - rowCount * TR_HEIGHT;
+    availableScrollY = Math.max(0, availableScrollY);
+
+    const newScrollPosition = this._clampScrollPosition(
+      this.lastScrollPosition + deltaY,
+      { min: 0, max: availableScrollY }
+    );
+
+    return {
+      totalHeight,
+      availableScrollY,
+      newScrollPosition,
+    };
+  }
+
+  /**
+   * Clamp scroll position within bounds
+   */
+  private _clampScrollPosition(value: number, bounds: ScrollBounds): number {
+    return Math.max(bounds.min, Math.min(value, bounds.max));
+  }
+
+  /**
+   * Calculate offset from scroll position
+   */
+  private _calculateOffsetFromScroll(scrollPosition: number): number {
+    return Math.ceil(scrollPosition / TR_HEIGHT);
+  }
+
+  /**
+   * Calculate touch scroll offset
+   */
+  private _calculateTouchScrollOffset(
+    deltaY: number,
+    touchStartOffset: number
+  ): number {
+    const rowCount = storeManager.getData('rowCount') as number;
+    const numberOfRecords = storeManager.getData('numberOfRecords') as number;
+
+    const availableHeight = rowCount * TR_HEIGHT;
+    const offsetRate = deltaY / availableHeight;
+
+    return Math.ceil(offsetRate * numberOfRecords) + touchStartOffset;
+  }
+
+  /**
+   * Clamp offset to valid range
+   */
+  private _clampOffsetToValidRange(offset: number): number {
+    const rowCount = storeManager.getData('rowCount') as number;
+    const numberOfRecords = storeManager.getData('numberOfRecords') as number;
+
+    const minOffset = 0;
+    const maxOffset = Math.max(0, numberOfRecords - rowCount);
+
+    return Math.max(minOffset, Math.min(offset, maxOffset));
+  }
+
+  /**
+   * Calculate scrollbar position and size
+   */
+  private _calculateScrollBarPosition(offset: number): ScrollBarCalculation {
+    const rowCount = storeManager.getData('rowCount') as number;
+    const numberOfRecords = storeManager.getData('numberOfRecords') as number;
+
+    const totalHeight = numberOfRecords * TR_HEIGHT;
+    const displayHeight = rowCount * TR_HEIGHT;
+    const displayRate = displayHeight / totalHeight;
+
+    let barHeight = Math.ceil(displayHeight * displayRate);
+    barHeight = Math.max(barHeight, MIN_SCROLLBAR_HEIGHT);
+
+    const availableHeight = displayHeight - barHeight;
+    const availableRate = availableHeight / totalHeight;
+    const barTop = Math.ceil(offset * TR_HEIGHT * availableRate);
+
+    return {
+      barHeight,
+      barTop,
+      displayRate,
+    };
+  }
+
+  /**
+   * Apply scrollbar styles
+   */
+  private _applyScrollBarStyles(
+    scrollBar: HTMLElement,
+    calculation: ScrollBarCalculation,
+    offset: number
+  ): void {
+    // Bar styles
+    scrollBar.style.height = `${calculation.barHeight}px`;
+    scrollBar.style.top = `${calculation.barTop}px`;
+
+    // Update position display
+    this._updatePositionDisplay(scrollBar, offset);
+
+    // Maintain active state
+    this.container.classList.add('-active');
+  }
+
+  /**
+   * Update position display
+   */
+  private _updatePositionDisplay(bar: HTMLElement, offset: number): void {
+    const position = bar.querySelector('.position') as HTMLElement;
+    if (position) {
+      position.textContent = String(offset + 1);
+    }
   }
 
   // ================================================================
